Fold the Omron WLP framing into a single OmronLib file.
